MULE-15403 - Forbidden flow names must tell me which one is triggering the exception
