+++ integration/src/test/java/org/mule/test/integration/AbstractConfigurationFailuresTestCase.java
@@ -1,0 +1,107 @@
+/*
+ * Copyright (c) MuleSoft, Inc.  All rights reserved.  http://www.mulesoft.com
+ * The software in this package is published under the terms of the CPAL v1.0
+ * license, a copy of which has been included with this distribution in the
+ * LICENSE.txt file.
+ */
+package org.mule.test.integration;
+
+import static java.lang.Thread.currentThread;
+import static java.util.Collections.emptyList;
+import static java.util.concurrent.TimeUnit.SECONDS;
+import static org.mule.runtime.config.api.SpringXmlConfigurationBuilderFactory.createConfigurationBuilder;
+import static org.mule.runtime.core.api.config.bootstrap.ArtifactType.APP;
+import static org.mule.runtime.core.api.context.notification.MuleContextNotification.CONTEXT_STARTED;
+import static org.mule.runtime.module.extension.api.loader.AbstractJavaExtensionModelLoader.TYPE_PROPERTY_NAME;
+import static org.mule.runtime.module.extension.api.loader.AbstractJavaExtensionModelLoader.VERSION;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.atomic.AtomicReference;
+
+import javax.inject.Inject;
+
+import org.mule.runtime.api.dsl.DslResolvingContext;
+import org.mule.runtime.api.exception.MuleException;
+import org.mule.runtime.api.meta.model.ExtensionModel;
+import org.mule.runtime.api.notification.IntegerAction;
+import org.mule.runtime.api.notification.NotificationListenerRegistry;
+import org.mule.runtime.api.util.concurrent.Latch;
+import org.mule.runtime.core.api.MuleContext;
+import org.mule.runtime.core.api.config.ConfigurationBuilder;
+import org.mule.runtime.core.api.config.DefaultMuleConfiguration;
+import org.mule.runtime.core.api.config.builders.AbstractConfigurationBuilder;
+import org.mule.runtime.core.api.context.DefaultMuleContextFactory;
+import org.mule.runtime.core.api.context.MuleContextBuilder;
+import org.mule.runtime.core.api.context.MuleContextFactory;
+import org.mule.runtime.core.api.context.notification.MuleContextNotification;
+import org.mule.runtime.core.api.context.notification.MuleContextNotificationListener;
+import org.mule.runtime.module.extension.api.loader.java.DefaultJavaExtensionModelLoader;
+import org.mule.runtime.module.extension.internal.manager.DefaultExtensionManager;
+import org.mule.tck.config.TestServicesConfigurationBuilder;
+import org.mule.tck.junit4.AbstractMuleTestCase;
+import org.mule.test.integration.exceptions.ErrorHandlingConfigurationFailuresTestCase;
+
+public class AbstractConfigurationFailuresTestCase extends AbstractMuleTestCase {
+
+  @Inject
+  private NotificationListenerRegistry notificationListenerRegistry;
+
+  protected void loadConfiguration(String configuration) throws MuleException, InterruptedException {
+
+    MuleContextFactory muleContextFactory = new DefaultMuleContextFactory();
+    List<ConfigurationBuilder> builders = new ArrayList<>();
+    builders.add(new AbstractConfigurationBuilder() {
+
+      @Override
+      protected void doConfigure(MuleContext muleContext) throws Exception {
+        DefaultExtensionManager defaultExtensionManager = new DefaultExtensionManager();
+        defaultExtensionManager.setMuleContext(muleContext);
+        defaultExtensionManager.initialise();
+        getRequiredExtensions().forEach(defaultExtensionManager::registerExtension);
+        muleContext.setExtensionManager(defaultExtensionManager);
+      }
+    });
+    builders.add(createConfigurationBuilder(configuration));
+    builders.add(new TestServicesConfigurationBuilder());
+    MuleContextBuilder contextBuilder = MuleContextBuilder.builder(APP);
+    final DefaultMuleConfiguration muleConfiguration = new DefaultMuleConfiguration();
+    muleConfiguration.setId(ErrorHandlingConfigurationFailuresTestCase.class.getSimpleName());
+    contextBuilder.setMuleConfiguration(muleConfiguration);
+    MuleContext muleContext = muleContextFactory.createMuleContext(builders, contextBuilder);
+    final AtomicReference<Latch> contextStartedLatch = new AtomicReference<>();
+    contextStartedLatch.set(new Latch());
+    notificationListenerRegistry.registerListener(new MuleContextNotificationListener<MuleContextNotification>() {
+
+      @Override
+      public boolean isBlocking() {
+        return false;
+      }
+
+      @Override
+      public void onNotification(MuleContextNotification notification) {
+        if (new IntegerAction(CONTEXT_STARTED).equals(notification.getAction())) {
+          contextStartedLatch.get().countDown();
+        }
+      }
+    });
+    muleContext.start();
+    contextStartedLatch.get().await(20, SECONDS);
+  }
+
+  protected List<ExtensionModel> getRequiredExtensions() {
+    return emptyList();
+  }
+
+  protected ExtensionModel loadExtension(Class extension, Set<ExtensionModel> deps) {
+    DefaultJavaExtensionModelLoader loader = new DefaultJavaExtensionModelLoader();
+    Map<String, Object> ctx = new HashMap<>();
+    ctx.put(TYPE_PROPERTY_NAME, extension.getName());
+    ctx.put(VERSION, "1.0.0-SNAPSHOT");
+    return loader.loadExtensionModel(currentThread().getContextClassLoader(), DslResolvingContext.getDefault(deps), ctx);
+  }
+
+}
