+++ integration/src/test/resources/org/mule/test/integration/validation/invalid-flow-name-config.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<mule xmlns="http://www.mulesoft.org/schema/mule/core"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xsi:schemaLocation="
+       http://www.mulesoft.org/schema/mule/core http://www.mulesoft.org/schema/mule/core/current/mule.xsd">
+
+    <flow name="flow/myFlow">
+        <logger/>
+        <error-handler>
+            <on-error-continue/>
+            <on-error-continue/>
+        </error-handler>
+    </flow>
+
+</mule>
